backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
aioodbc==0.5.0
sqlalchemy[asyncio]==2.0.30
pyodbc==5.1.0                 # ODBC driver
pydantic-settings==2.2.1
# paho-mqtt==2.1.0
paho-mqtt==1.6.1
async-timeout==4.0.3
python-dotenv==1.0.1
pytest==8.3.5
pytest-asyncio==0.26.0
ipaddress==1.0.23
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

aiomqtt==2.4.0
# asyncio_mqtt==0.16.2
aiofiles==23.2.1
openpyxl==3.1.5 
pandas==2.3.2